Tos attempted to create a power relay event for me
Description by Tos:
You put a "Destroyable" bool-value and a "Health" int-value inside the model
I have modified the "DamageFunc" function and line 102 + line 143 (I passed the player once calling the damage function) inside "ServerAbilityManager", which will check if it's an object and will damage it. You would have to pass the player too, because I am using it to figure out the stage level. If the stage level is 3 or bigger, it will do damage.